tests/wpt/tests/websockets/interfaces/WebSocket/url/resolve.html

<!doctype html>
<title>WebSocket#url: resolving</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.sub.js></script>
<meta name="variant" content="?default">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
test(function() {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT + '/echo?foo%20bar baz');
  assert_equals(ws.url, SCHEME_DOMAIN_PORT + '/echo?foo%20bar%20baz');
});
</script>
